requirements.txt
opencv-python-headless==4.9.0.80
pillow==8.4.0